move some react components into isolated folders
